feat(backend): update endpoint for getting files paginated to use page instead of custom result

+++ ScriptBeeWebApp/src/Gateway/Adapters/Web/EndpointDefinitions/ProjectStructure/Contracts/WebGetProjectFilesResponse.cs
@@ -2,7 +2,7 @@ namespace ScriptBee.Web.EndpointDefinitions.ProjectStructure.Contracts;
 
 public sealed record WebGetProjectFilesResponse(
     IEnumerable<WebProjectFileNode> Data,
-    int TotalCount,
+    long TotalCount,
     int Offset,
     int Limit
 )
